refactor: extract magic numbers, quality/safety improvements
Phase 3 — Named constants:
- Created mshw0231-raw-constants.h with 19 constants:
  signal thresholds, peak detection, CCL, blob splitting, edge penalty,
  Hungarian costs, slot state machine, association radius multipliers
- Replaced all hardcoded numbers across the pipeline

Phase 4 — Quality/safety:
- c590 LUT: round-half-up for better precision (+ C590_STEP_DEN/2)
- Updated stale comment (288-col → 72x48 grid)
- BFS queue type: u16 → u32 (prevent truncation for large grids)
- Removed 4 redundant heatmap_signal[nxt] > 0 checks in BFS
- Added touch_input null guard at process_samples entry

+++ driver/spi-hid-core.h
@@ -236,7 +236,7 @@ struct spi_hid {
 	u8  heatmap_touched[HEATMAP_MAX_CELLS];
 	s16 heatmap_signal[HEATMAP_MAX_CELLS];   /* precomputed c590 signal rise */
 	u16 heatmap_label[HEATMAP_MAX_CELLS];    /* CCL component labels */
-	u16 heatmap_queue[HEATMAP_MAX_CELLS];    /* CCL flood-fill queue */
+	u32 heatmap_queue[HEATMAP_MAX_CELLS];    /* CCL flood-fill queue */
 
 	/* Blob state. Coordinates are fixed-point grid ×100. */
 	u32 blob_x[HEATMAP_MAX_BLOBS];
